moving legacy rnn model
